requirements.txt
appdirs==1.4.4
asgiref==3.3.1
beautifulsoup4==4.9.3
black==20.8b1
boto3==1.17.52
botocore==1.20.52
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
cssutils==2.3.0
dj-database-url==0.5.0
Django==3.1.5
django-modeladmin-reorder==0.3.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
django-storages==1.11.1
gunicorn==20.1.0
idna==2.10
importlib-metadata==6.0.0
jmespath==0.10.0
mypy-extensions==0.4.3
pathspec==0.8.1
Pillow==9.5.0
psycopg2-binary==2.9.5
python-dateutil==2.8.1
pytz==2020.5
regex==2020.11.13
requests==2.25.1
s3transfer==0.3.7
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
whitenoise==5.2.0
zipp==3.13.0
